place lambda function on correct directory
