+++ motor-on-voice/lambda_function.py
@@ -1,6 +1,7 @@
-from flask import Flask, render_template
+from flask import Flask, render_template, json
 from flask_ask import Ask, statement, question, session
 import json
+import requests
 
 app = Flask(__name__)
 ask = Ask(app, '/')
@@ -9,16 +10,19 @@ motor_status = 0
 def lambda_handler(event, _context):
     return ask.run_aws_lambda(event)
 
+# will run on invocation
 @ask.launch
 def launched():
    return question('Hello, How can I help you today')
 
+# when something goes wrong
 @ask.default_intent
 def default_message():
     speech = "Oh-aw! I couldn't understand you. Do you need any assistant?"
     retry_speech = "Please try again later. I couldn't rectify your issue"
     return statement(speech).reprompt(retry_speech)
 
+# start motor if not running
 @ask.intent('MotorOnIntent')
 def start_motor():
     global motor_status
@@ -30,6 +34,7 @@ def start_motor():
         motor_status = 1
     return statement(response)
 
+# stop motor if running
 @ask.intent('MotorOffIntent')
 def stop_motor():
     global motor_status
@@ -41,6 +46,7 @@ def stop_motor():
         motor_status = 0
     return statement(response)
 
+# return whether motor runnning or not
 @ask.intent('MotorStatusIntent')
 def status_motor():
     global motor_status
@@ -53,6 +59,7 @@ def status_motor():
         session.attributes['status'] = "start"
     return question(response)
 
+# start or stop motor as a confirmation
 @ask.intent('YesIntent')
 def restart_motor():
     global motor_status
@@ -65,10 +72,30 @@ def restart_motor():
         motor_status = 0
     return statement(response)
 
+# don't start or stop motor now
 @ask.intent('NoIntent')
 def motor_do_nothing():
     response = render_template('motor_do_nothing')
     return statement(response)
+
+# HTTP GET req for motor status
+@app.route('/motor/status')
+def get_data():
+    global motor_status
+    response = ''
+    if(motor_status == 1):
+        response = app.response_class(
+            response="running",
+            status=200,
+            mimetype='application/json'
+        )
+    else:
+        response = app.response_class(
+            response="stopped",
+            status=200,
+            mimetype='application/json'
+        )
+    return response
 
 if __name__ == '__main__':
     app.run(debug=True)
